requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
redis==5.0.8
tenacity==9.0.0
httpx==0.27.2
requests==2.32.3
python-slugify==8.0.4
pytrends==4.9.2
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
loguru==0.7.2
